adapted ProductDataset and moved to src/export
